+++ tasks/voc-py35/tools/run.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
-echo "Testing $GITHUB_OWNER/$GITHUB_PROJECT_NAME @ $SHA"
 echo "Python version=`python --version`"
 java -version
 ant -version
 echo
-# Download and unpack code at the test SHA
-echo "curl -L $CODE_URL -o code.zip"
-curl -s -L $CODE_URL -o code.zip
-echo "--------------------------------------------------------------------------------"
-echo "unzip code.zip"
-unzip code.zip
-cd $GITHUB_PROJECT_NAME-$SHA
+if [ -e "/app/beekeeper.yml" ]; then
+    echo "Testing local code"
+else
+    echo "Testing $GITHUB_OWNER/$GITHUB_PROJECT_NAME @ $SHA"
+    echo
+    # Download and unpack code at the test SHA
+    echo "curl -L $CODE_URL -o code.zip"
+    curl -s -L $CODE_URL -o code.zip
+    echo "--------------------------------------------------------------------------------"
+    echo "unzip code.zip"
+    unzip code.zip
+    cd $GITHUB_PROJECT_NAME-$SHA
+fi
 echo "--------------------------------------------------------------------------------"
 echo pip install -e .
 pip install -e .
